+++ src/main/java/com/github/shynixn/blockball/event/BallEvent.kt
@@ -2,7 +2,6 @@ package com.github.shynixn.blockball.event
 
 import com.github.shynixn.blockball.contract.SoccerBall
 
-
 /**
  * Base Event for all SoccerBall events.
  */
@@ -11,4 +10,4 @@ open class BallEvent(
      * SoccerBall.
      */
     val ball: SoccerBall
-) : BlockBallEvent()
+) : BlockBallEvent(false)