+++ src/main/java/com/github/shynixn/blockball/event/BallRayTraceEvent.kt
@@ -1,8 +1,8 @@
 package com.github.shynixn.blockball.event
 
 import com.github.shynixn.blockball.contract.SoccerBall
-import com.github.shynixn.mcutils.packet.api.meta.enumeration.BlockDirection
 import org.bukkit.Location
+import org.bukkit.block.BlockFace
 
 /**
  * The ray trace event is called when the ball has got a velocity in a certain
@@ -19,17 +19,17 @@ class BallRayTraceEvent(
     /**
      * Modifies if the raytrace has ended in a block hit.
      */
-    var hitBlock: Boolean = false,
+    var hasHitBlock: Boolean = false,
 
     /**
      * Gets the resulting position if the object actually
      * performs the raytrace. Current Position is the position of the ball.
      */
-    val targetLocation: Location,
+    var targetLocation: Location,
 
     /**
      * Block Direction which was hit. If hitBlock is false, it may contain any
      * direction. If changed and the hitBlock is true the outgoing Vector is differently calculated later.
      */
-    var blockDirection: BlockDirection
+    var blockFace: BlockFace? = null
 ) : BallEvent(ball)