+++ src/main/java/com/github/shynixn/blockball/event/BallRemoveEvent.kt
@@ -2,7 +2,6 @@ package com.github.shynixn.blockball.event
 
 import com.github.shynixn.blockball.contract.SoccerBall
 
-
 /**
  * Event which gets sent when the ball is requested to get removed.
  */