+++ src/main/java/com/github/shynixn/blockball/impl/listener/BallListener.kt
@@ -1,97 +1,50 @@
 package com.github.shynixn.blockball.impl.listener
 
-import com.github.shynixn.blockball.contract.SoccerBall
-import com.github.shynixn.blockball.contract.SoccerBallFactory
-import com.github.shynixn.blockball.enumeration.BallActionType
-import com.github.shynixn.blockball.event.*
-import com.github.shynixn.shyparticles.contract.ParticleEffectService
+import com.github.shynixn.blockball.contract.SoccerBallService
+import com.github.shynixn.blockball.enumeration.BallInputActionType
+import com.github.shynixn.mccoroutine.folia.launch
+import com.github.shynixn.mccoroutine.folia.regionDispatcher
+import com.github.shynixn.mcutils.packet.api.event.PacketAsyncEvent
+import com.github.shynixn.mcutils.packet.api.meta.enumeration.InteractionType
+import com.github.shynixn.mcutils.packet.api.packet.PacketInInteractEntity
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
+import org.bukkit.event.block.Action
+import org.bukkit.event.player.PlayerInteractEvent
+import org.bukkit.plugin.Plugin
 
-/**
- * Handles common ball events.
- */
-class BallListener(
-    private val soccerBallFactory: SoccerBallFactory,
-    private val effectService: ParticleEffectService
-) : Listener {
+class BallListener(private val plugin: Plugin, private val soccerBallService: SoccerBallService) : Listener {
     /**
-     * Gets called when the ball raytraces in the world.
+     * Gets called when a packet arrives.
      */
     @EventHandler
-    fun ballRayTraceEvent(event: BallRayTraceEvent) {
-        this.playEffects(event.ball, BallActionType.ONMOVE)
-    }
+    fun onPacketEvent(event: PacketAsyncEvent) {
+        val packet = event.packet
 
-    /**
-     * Gets called when a ball dies.
-     *
-     * @param event event
-     */
-    @EventHandler
-    fun ballDeathEvent(event: BallRemoveEvent) {
-        this.soccerBallFactory.removeTrackedBall(event.ball)
-    }
+        if (packet !is PacketInInteractEntity) {
+            return
+        }
 
-    /**
-     * Gets called when a player left clicks a ball.
-     *
-     * @param event event
-     */
-    @EventHandler
-    fun ballKickEvent(event: BallLeftClickEvent) {
-        this.playEffects(event.ball, BallActionType.ONKICK)
-    }
-
-    /**
-     * Gets called when a player right clicks a ball.
-     */
-    @EventHandler
-    fun ballPassEvent(event: BallRightClickEvent) {
-        this.playEffects(event.ball, BallActionType.ONPASS)
-    }
-
-    /**
-     * Gets called when a player interacts a ball.
-     *
-     * @param event event
-     */
-    @EventHandler
-    fun ballInteractEvent(event: BallTouchPlayerEvent) {
-        this.playEffects(event.ball, BallActionType.ONINTERACTION)
-    }
-
-    /**
-     * Gets called when the ball spawns.
-     *
-     * @param event event
-     */
-    @EventHandler
-    fun ballSpawnEvent(event: BallSpawnEvent) {
-        this.playEffects(event.ball, BallActionType.ONSPAWN)
-    }
-
-    /**
-     * Gets called when a ball gets shot into goal.
-     *
-     * @param event event
-     */
-    @EventHandler
-    fun gameGoalEvent(event: GameGoalEvent) {
-        playEffects(event.game.ball!!, BallActionType.ONGOAL)
-    }
-
-    /**
-     * Plays effects.
-     */
-    private fun playEffects(ball: SoccerBall, actionEffect: BallActionType) {
-        val effectName = ball.meta.effects[actionEffect]
-
-        if (effectName != null) {
-            val effect = effectService.getEffectMetaFromName(effectName)
-            if (effect != null) {
-                effectService.startEffect(effect, { ball.getLocation() }, null, null)
+        val ball = soccerBallService.getByEntityId(packet.entityId) ?: return
+        plugin.launch(plugin.regionDispatcher(ball.getLocation())) {
+            if (packet.actionType == InteractionType.ATTACK) {
+                ball.applyInteraction(event.player, BallInputActionType.LEFT_CLICK)
+            } else {
+                ball.applyInteraction(event.player, BallInputActionType.RIGHT_CLICK)
             }
+        }
+    }
+
+    /**
+     * Gets called when ball is already grabbed.
+     */
+    @EventHandler
+    fun onPlayerInteractEvent(event: PlayerInteractEvent) {
+        val ball = soccerBallService.getAll().firstOrNull { e -> e.grabbingPlayer == event.player }
+        if (event.action == Action.LEFT_CLICK_AIR || event.action == Action.LEFT_CLICK_BLOCK) {
+            ball?.applyInteraction(event.player, BallInputActionType.LEFT_CLICK)
+        } else if (event.action == Action.RIGHT_CLICK_BLOCK || event.action == Action.RIGHT_CLICK_AIR) {
+            ball?.applyInteraction(event.player, BallInputActionType.RIGHT_CLICK)
         }
     }
 }